Tons of changes to fQuery to hopefully driver-ize it.
Likely going to have to convert the format from string to SQL to string to array or something, otherwise it just won't be able to be driverized (unless we convert SQL statements which seems vastly more inefficient.

Just going to take the pre-parseSQL array in fQuery and that is my new driver format!